front: move nativeEvents and nativeInfos in dedicated folder (#1027)
