+++ public/mobile-app/src/lib/bridges/nativeBridge.ts
