+++ public/mobile-app/src/lib/bridges/nativeInfos.ts
