+++ public/mobile-app/src/lib/state/User.svelte.ts
@@ -9,8 +9,8 @@ import { Address } from '$lib/address';
 import { callBAN } from '$lib/addressesFromBAN';
 import type { APIAgendaItem } from '$lib/api-agenda';
 import * as auth from '$lib/auth';
+import { emit } from '$lib/bridges/nativeBridge';
 import { franceConnectLogout, parseJwt } from '$lib/france-connect';
-import { emit } from '$lib/nativeEvents';
 import { disableNotifications } from '$lib/notifications';
 import { Preferences, type ZoneInfo } from '$lib/state/preferences';
 import { formatDate } from '$lib/utils';
